Move Python utilities from `circuit` to `util`
The base definitions of these objects are unrelated to `qiskit-circuit`
and the core circuit IR concepts, and can easily be moved into the
`qiskit-util` crate to break further spurious dependencies on the IR
definitions.
