+++ crates/circuit/Cargo.toml
@@ -13,7 +13,7 @@ doctest = false
 workspace = true
 
 [dependencies]
-qiskit-util.workspace = true
+qiskit-util = { workspace = true, features = ["py"] }
 rayon.workspace = true
 ahash.workspace = true
 rustworkx-core.workspace = true
