+++ crates/quantum_info/Cargo.toml
@@ -11,7 +11,7 @@ doctest = false
 
 [dependencies]
 qiskit-circuit.workspace = true
-qiskit-util.workspace = true
+qiskit-util = { workspace = true, features = ["py"] }
 numpy.workspace = true
 num-complex.workspace = true
 num-traits.workspace = true
