+++ crates/quantum_info/src/lib.rs
@@ -25,7 +25,7 @@ mod test;
 use pyo3::import_exception;
 
 pub(crate) mod imports {
-    use qiskit_circuit::imports::ImportOnceCell;
+    use qiskit_util::py::ImportOnceCell;
 
     pub static PAULI_TYPE: ImportOnceCell = ImportOnceCell::new("qiskit.quantum_info", "Pauli");
     pub static PAULI_LIST_TYPE: ImportOnceCell =
