+++ crates/util/Cargo.toml
@@ -9,8 +9,13 @@ license.workspace = true
 name = "qiskit_util"
 doctest = false
 
+[features]
+py = ["dep:pyo3"]
+
 [lints]
 workspace = true
 
 [dependencies]
 num-complex.workspace = true
+pyo3 = { workspace = true, optional = true }
+thiserror.workspace = true
